requirements.txt
imagehash
ipywidgets
jupyterlab
nodejs
notebook
pandas
pillow
pydicom
seaborn
tqdm
yacs